Updated index.html and map.js
Removed the infobox plugin script from index.html, since I think
InfoWindow is a little easier to use. Also added a footer with a back
button that will eventually be used to go back to the State screen.
Changed screenWidth to use screen.width instead of JQuery
$(window).width().
Added a variable called centerOfUSA, easier to tell what that is.
Added code to keep the center of the USA inside a bounding rectangle, so
the user can't just pan around the whole world.