Gets Latitude/Longitude of each state from server
Edited the index file to add a screen where you can pick each state from
a list (since we don't have anything to scan). Once on the state screen,
user can tap "Map" button to go to that state's map. On "pageshow" of
the #map screen, ajax gets the states from the server, loops through to
get the current state's lat/lon, and saves that as a variable, which is
used for an InfoWindow pointing to that state. The InfoWindow can have
the state artwork in it.